result: make Result RunInfo
Rename it to RunInfo and make it an abstract class.
Later, we will have implementations of this class
that will either get the values from xml or from database,
so in the later case we do not need to copy all the information
again, but keep the object returned from query
and make queries upon it.

Signed-off-by: Marek Chalupa <xchalup4@fi.muni.cz>