Store category for each benchmark separately
Previously, only one category was stored for the *whole* test run.

Not a particularly efficient solution due to added redundancy, but it does
the job.

Fixes issue when a log for different category was tried to be opened (e.g.
`no-overflow` for `unreach-call` benchmark):

"There is no item named 'archive/SV-COMP22_no-overflow.test.yml.log' in the
archive"

when the file prefix should have been `SV-COMP22_unreach-call` instead.